Resolve minor mistake

+++ classes/Menu.cpp
@@ -9,7 +9,7 @@ Menu::Menu(string input, vector<string> menuOptions){
 }
 
 void Menu::printMenu(){
-    for(int i = 0; i < menuOptions.size(); i++){
+    for(size_t i = 0; i < menuOptions.size(); i++){
         cout << i << '-' << menuOptions.at(i) << '\n';
     }
     cout << inputPrompt;
